requirements.txt
Django==2.2
djangorestframework==3.12.4
google-api-python-client==2.9.0
celery==5.1.1
django-celery-results==2.1.0
SQLAlchemy==1.4.18
